Removed excessive Now playing messages
Prior to this the recursive music playing function would name each song it was now playing. Now it only shows it for the first song played and for none in the queue. Playl will not show any song name at all but say that the playlist is loaded when it is finished.